Version 0.6 Fully functional Session-Identification System. Added restriction in user rights

+++ index.php
@@ -2,7 +2,7 @@
 
 /*
 controller for a lending system
-version 0.5
+version 0.6
 date 25.10.18
  */
 session_start();
@@ -36,8 +36,9 @@ elseif ((!isset ($oObject->r_ac)) or ($sName == 'logi') or ($sName == 'strt') or
 	$oObject = new Data;
 }
 //view header
-include ('views/header.php');
-
+if ($oObject->r_ac!="book_show_plain"){
+	include ('views/header.php');
+}
 //methods
 switch ($oObject->r_ac){
 	case 'strt':
@@ -53,70 +54,138 @@ switch ($oObject->r_ac){
 		break;
 		
 	case 'book_new':
+		if ($_SESSION['admin']==1){	
 		include ('views/book_form.php');
+		}
+		else{
+			echo "Keine Berechtigung";
+		}
 		break;
 	
 	case 'book_change':
+		if ($_SESSION['admin']==1){	
 		$oObject->aRow_all = $oObject->get_book();
 		$oObject->aRow = $oObject->aRow_all[$oObject->r_book_ID];
 		include ('views/book_form.php');
+		}
+		else{
+			echo "Keine Berechtigung";
+		}
 		break;	
 	case 'book_save':
+		if ($_SESSION['admin']==1){	
 		$oObject->save_book();
 		$oObject->r_book_ID = NULL;
 		$oObject->aBook = $oObject->get_book();
 		include ("views/all_books.php");
+		}
+		else{
+			echo "Keine Berechtigung";
+		}
 		break;
 	case 'book_show':
-		$oObject->set_session();
+		$oObject->aBook = $oObject->get_book();
+		include ("views/all_books.php");
+		break;
+	case 'book_show_plain':
 		$oObject->aBook = $oObject->get_book();
 		include ("views/all_books.php");
 		break;
 	case 'book_delete':
+		if ($_SESSION['admin']==1){	
 		$oObject->delete_book();
 		$oObject->r_book_ID = NULL;
 		$oObject->aBook = $oObject->get_book();
 		include ("views/all_books.php");
+		}
+		else{
+			echo "Keine Berechtigung";
+		}
 		break;
 	case 'user_new':
+		if ($_SESSION['admin']==1){	
 		include ("views/user_form.php");
+		}
+		else{
+			echo "Keine Berechtigung";
+		}
 		break;
 	case 'user_save':
+		if (($_SESSION['admin']==1) or ($_SESSION['user_ID']==$oObject->r_user_ID)){	
 		$oObject->save_user();
 		$oObject->r_user_ID = NULL;
 		$oObject->aUser = $oObject->get_user();
 		include ("views/all_user.php");
+		}
+		else{
+			echo "Keine Berechtigung";
+		}
 		break;
 	case 'user_delete':
+		if (($_SESSION['admin']==1) or ($_SESSION['user_ID']==$oObject->r_user_ID)){	
 		$oObject->delete_user();
 		$oObject->r_user_ID = NULL;
 		$oObject->aUser = $oObject->get_user();
 		include ("views/all_user.php");
+		}
+		else{
+			echo "Keine Berechtigung";
+		}
 		break;
-	case 'user_show':
+	case 'user_self':
+		$oObject->r_user_ID =$_SESSION['user_ID'];
 		$oObject->aUser = $oObject->get_user();
 		include ("views/all_user.php");
 		break;
+	case 'user_show':
+		if ($_SESSION['admin']==1){	
+		$oObject->aUser = $oObject->get_user();
+		include ("views/all_user.php");
+		}
+		else{
+			echo "Keine Berechtigung";
+		}
+		break;
 	case 'user_search':	
+		if ($_SESSION['admin']==1){	
 		include ("views/user_search.php");
+		}
+		else {
+			echo "Keine Berechtigung";
+		}
 		break;
 	case 'user_change':
-	
+		if (($_SESSION['admin']==1) or ($_SESSION['user_ID']==$oObject->r_user_ID)){	
 		$oObject->aRow_all = $oObject->get_user();
 		$oObject->aRow = $oObject->aRow_all[$oObject->r_user_ID];
 		include ("views/user_form.php");
+		}
+		else{
+			echo "Keine Berechtigung";
+		}
 		break;
 		
 	case 'lend_new':
+		if ($_SESSION['admin']==1){	
 		include ("views/lend_form.php");
+		}
+		else{
+			echo "Keine Berechtigung";
+		}
 		break;
 	case 'lend_save':
+		if ($_SESSION['admin']==1){	
 		$oObject->save_lend();
 		$oObject->r_lend_ID = NULL;
 		$oObject->aLend = $oObject->get_lend();
 		include ("views/all_lend.php");
+		}
+		else{
+			echo "Keine Berechtigung";
+		}
 		break;
 	case 'lend_return':
+		if ($_SESSION['admin']==1){	
 		$oObject->return_lend();
 	//	$oBook = new Book();
 	//	$oBook->return_book($oObject->r_book_ID);
@@ -124,24 +193,46 @@ switch ($oObject->r_ac){
 		$oObject->r_book_ID = NULL;
 		$oObject->aLend = $oObject->get_lend();
 		include ("views/all_lend.php");
+		}
+		else{
+			echo "Keine Berechtigung";
+		}
 		break;
 	case 'lend_delete':
+		if ($_SESSION['admin']==1){	
 		$oObject->delete_lend();
 	//	$oBook = new Book();
 	//	$oBook->return_book($oObject->r_book_ID);
 		$oObject->r_lend_ID = NULL;
 		$oObject->aLend = $oObject->get_lend();
 		include ("views/all_lend.php");
+		}
+		else{
+			echo "Keine Berechtigung";
+		}
 		break;
 	case 'lend_show':
-
+		if (($_SESSION['admin']==1) or ($_SESSION['user_ID'] == $oObject->r_user_ID)){	
 		$oObject->aLend = $oObject->get_lend();
-
+		include ("views/all_lend.php");
+		}
+		else{
+			echo "Keine Berechtigung";
+		}
+		break;
+	case 'lend_self':
+		$oObject->r_user_ID = $_SESSION['user_ID'];
+		$oObject->aLend = $oObject->get_lend();
 		include ("views/all_lend.php");
 		break;
 	case 'lend_change':
+		if ($_SESSION['admin']==1){	
 		$oObject->get_lend();
 		include ("views/lend_form.php");
+		}
+		else{
+			echo "Keine Berechtigung";
+		}
 		break;
 	default: 
 		$oObject->aBook = $oObject->get_book();
